+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-16 16:26:53
-// Commit: fix(test): correct layout constraints and overflow in test
+// Updated at: 2026-04-16 18:09:10
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '429';
-  final String lastMessage = 'fix(test): correct layout constraints and overflow in test';
-  final DateTime timestamp = DateTime.parse('2026-04-16T16:26:53+05:00');
+  final String commitRef = '436';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-04-16T18:09:10+05:00');
 }